+++ sat/cli/bootsys/cabinet_power.py
@@ -1,7 +1,7 @@
 """
 Powers on and off liquid-cooled compute cabinets.
 
-(C) Copyright 2020-2021 Hewlett Packard Enterprise Development LP.
+(C) Copyright 2020-2022 Hewlett Packard Enterprise Development LP.
 
 Permission is hereby granted, free of charge, to any person obtaining a
 copy of this software and associated documentation files (the "Software"),
@@ -24,9 +24,9 @@ OTHER DEALINGS IN THE SOFTWARE.
 import logging
 
 from sat.apiclient import APIError, CAPMCClient, HSMClient
-from sat.cli.bootsys.discovery import HMSDiscoveryCronJob, HMSDiscoveryError, HMSDiscoveryScheduledWaiter
 from sat.cli.bootsys.power import CAPMCPowerWaiter
 from sat.config import get_config_value
+from sat.hms_discovery import HMSDiscoveryCronJob, HMSDiscoveryError, HMSDiscoveryScheduledWaiter
 from sat.session import SATSession
 from sat.util import prompt_continue