+++ sat/cli/swap/main.py
@@ -1,7 +1,7 @@
 """
 The main entry point for the swap command.
 
-(C) Copyright 2020-2021 Hewlett Packard Enterprise Development LP.
+(C) Copyright 2020-2022 Hewlett Packard Enterprise Development LP.
 
 Permission is hereby granted, free of charge, to any person obtaining a
 copy of this software and associated documentation files (the "Software"),
@@ -24,10 +24,39 @@ OTHER DEALINGS IN THE SOFTWARE.
 
 import logging
 
-from sat.cli.swap.switch import swap_switch
+from sat.cli.swap.blade import swap_blade
 from sat.cli.swap.cable import swap_cable
+from sat.cli.swap.errors import ERR_INVALID_OPTIONS
+from sat.cli.swap.switch import swap_switch
+from sat.util import pester
+
 
 LOGGER = logging.getLogger(__name__)
+
+
+def check_arguments(component_type, action, dry_run, disruptive):
+    """Check that given options are valid for a swap.
+
+    Args:
+        component_type (str): 'blade', 'switch', or 'cable'
+        action (str): 'enable' or 'disable'
+        dry_run (bool): if True, skip applying configuration to ports
+        disruptive (bool): if True, do not confirm disable/enable
+
+    Raises:
+        SystemExit(1): if invalid options
+    """
+    if not dry_run and not action:
+        LOGGER.error('The action option is required if not a dry run, exiting.')
+        raise SystemExit(ERR_INVALID_OPTIONS)
+
+    if dry_run and action:
+        LOGGER.error('The action option is not valid with the dry run option.')
+        raise SystemExit(ERR_INVALID_OPTIONS)
+
+    if not (disruptive or dry_run or
+            pester(f'Enable/disable of {component_type} can impact system. Continue?')):
+        raise SystemExit(ERR_INVALID_OPTIONS)
 
 
 def do_swap(args):
@@ -39,8 +68,11 @@ def do_swap(args):
     Returns:
         None
     """
+    check_arguments(args.target, args.action, args.dry_run, args.disruptive)
 
     if args.target == 'cable':
         swap_cable(args)
     elif args.target == 'switch':
         swap_switch(args)
+    elif args.target == 'blade':
+        swap_blade(args)