+++ CHANGELOG.md
@@ -49,6 +49,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ``--cfs-fields``.
 - Added a ``--bos-template`` option to ``sat status`` which allows the status
   report to be filtered by the specified session template's boot sets.
+- Added a ``sat swap blade`` subcommand which automates the procedure for
+  swapping compute and UAN blades.
 
 ### Changed
 - Changed the output of ``sat status`` to split different component types into